# specs/001-pilot-space-mvp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Pilot Space API
  description: |
    AI-augmented SDLC platform API for project management with intelligent features.

    ## Authentication
    All endpoints require Bearer token authentication via Supabase Auth (GoTrue).
    Authorization is enforced via Row-Level Security (RLS) policies.

    ## Rate Limits
    - Standard endpoints: 1000 requests/minute per workspace
    - AI endpoints: 100 requests/minute per workspace (subject to LLM provider limits)
    - Webhook endpoints: 500 requests/minute

    ## Pagination
    List endpoints support cursor-based pagination using `cursor` and `limit` parameters.
    Response format: `{ data: [...], meta: { total, cursor, hasMore } }`

    ## Updated
    Session 2026-01-22: Migrated to Supabase platform
  version: 2.0.0
  contact:
    name: Pilot Space Team
    url: https://pilot-space.com
  license:
    name: MIT
servers:
  - url: https://api.pilot-space.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Development
...
paths:
...
  /workspaces/{workspaceSlug}/projects:
    parameters:
      - $ref: "#/components/parameters/workspaceSlug"
    get:
      tags: [Projects]
      summary: List projects in workspace
      operationId: listProjects
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

    post:
      tags: [Projects]
      summary: Create project
      operationId: createProject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /workspaces/{workspaceSlug}/projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/workspaceSlug"
      - $ref: "#/components/parameters/projectId"
    get:
      tags: [Projects]
      summary: Get project
      operationId: getProject
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  # ============================================================================
  # Notes
  # ============================================================================
...
  /webhooks/github:
    post:
      tags: [Webhooks]
      summary: GitHub webhook handler
      operationId: handleGitHubWebhook
      security: []
      parameters:
        - name: X-GitHub-Event
          in: header
          required: true
          schema:
            type: string
        - name: X-Hub-Signature-256
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Webhook processed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    workspaceSlug:
      name: workspaceSlug
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9-]+$"
    projectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        identifier:
          type: string
        description:
          type: string
        emoji:
          type: string
        coverImageUrl:
          type: string
          format: uri
        isPublic:
          type: boolean
        issueCount:
          type: integer
        settings:
          type: object
        createdAt:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      required: [name, identifier]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        identifier:
          type: string
          pattern: "^[A-Z0-9]+$"
          minLength: 2
          maxLength: 10
        description:
          type: string
        emoji:
          type: string
        isPublic:
          type: boolean
          default: false

    # ==================== Note ====================
...